Revert "feat(server): implement session-based MCP architecture (#28)"
This reverts commit 33f947432854559befea4002f91956d9dd784c15.
