+++ packages/server/src/index.test.ts
@@ -12,7 +12,6 @@ const TEST_CONFIG: PolygConfig = {
 };
 
 import {
-  createMcpServer,
   formatToolError,
   HealthChecker,
   HTTPTransport,
@@ -22,11 +21,6 @@ import {
   ServerError,
   ServerStartError,
   ServerStopError,
-  SessionLimitError,
-  SessionManager,
-  SessionNotFoundError,
-  SessionRequiredError,
-  SharedResources,
   ToolExecutionError,
   ToolInputValidationError,
   TransportConfigError,
@@ -49,24 +43,6 @@ describe('server exports', () => {
 
   it('exports HealthChecker', () => {
     expect(HealthChecker).toBeDefined();
-  });
-
-  it('exports SharedResources', () => {
-    expect(SharedResources).toBeDefined();
-  });
-
-  it('exports SessionManager', () => {
-    expect(SessionManager).toBeDefined();
-  });
-
-  it('exports createMcpServer', () => {
-    expect(createMcpServer).toBeDefined();
-  });
-
-  it('exports session error types', () => {
-    expect(SessionNotFoundError).toBeDefined();
-    expect(SessionLimitError).toBeDefined();
-    expect(SessionRequiredError).toBeDefined();
   });
 });
 
@@ -109,9 +85,9 @@ describe('HTTPTransport', () => {
     expect(transport).toBeDefined();
   });
 
-  it('should throw if started without resources attached', async () => {
+  it('should throw if started without server attached', async () => {
     const transport = new HTTPTransport({ port: 3000 });
-    await expect(transport.start()).rejects.toThrow('No resources attached');
+    await expect(transport.start()).rejects.toThrow('No server attached');
   });
 
   it('should return null address before starting', () => {
