+++ packages/server/src/index.ts
@@ -3,14 +3,7 @@
 export * from './errors.js';
 export { HealthChecker } from './health.js';
 export { HTTPTransport } from './http.js';
-export { createMcpServer } from './mcp-server-factory.js';
 export { PolygMCPServer } from './server.js';
-export {
-  type SessionContext,
-  SessionManager,
-  type SessionManagerOptions,
-} from './session-manager.js';
-export { SharedResources } from './shared-resources.js';
 export * from './tools/index.js';
 
 export const VERSION = '0.1.0';
