+++ packages/server/src/server.test.ts
@@ -6,11 +6,6 @@ import { PolygMCPServer } from './server.js';
 // Set KEEP_TEST_DATA=1 to preserve test data for visualization
 const KEEP_TEST_DATA = process.env.KEEP_TEST_DATA === '1';
 
-// Helper to wait for FalkorDB eventual consistency
-// FalkorDB writes may not be immediately visible to subsequent queries
-const waitForConsistency = (ms = 500) =>
-  new Promise((resolve) => setTimeout(resolve, ms));
-
 // Test config with mock API key
 const TEST_CONFIG: PolygConfig = {
   ...DEFAULT_CONFIG,
@@ -229,22 +224,19 @@ describe('PolygMCPServer tool handlers', () => {
       expect(retrieved?.name).toBe(name);
     });
 
-    // TODO: These tests are flaky due to FalkorDB eventual consistency issues
-    // when running in parallel with other tests. Need to investigate test isolation.
-    it.skip('should retrieve an entity by name', async () => {
+    it('should retrieve an entity by name', async () => {
       const graphs = server.getOrchestrator().getGraphs();
       const uniqueId = Date.now().toString();
       const name = `Charlie_${uniqueId}`;
 
       await graphs.entity.addEntity(name, 'person');
-      await waitForConsistency();
       const retrieved = await graphs.entity.getEntity(name);
 
       expect(retrieved).not.toBeNull();
       expect(retrieved?.name).toBe(name);
     });
 
-    it.skip('should link two entities', async () => {
+    it('should link two entities', async () => {
       const graphs = server.getOrchestrator().getGraphs();
 
       // Use unique names to avoid conflicts with parallel test runs
@@ -257,9 +249,6 @@ describe('PolygMCPServer tool handlers', () => {
         `Bob_link_${uniqueId}`,
         'person',
       );
-
-      // Wait for FalkorDB eventual consistency before linking
-      await waitForConsistency();
 
       // Link them
       await graphs.entity.linkEntities(alice.uuid, bob.uuid, 'knows');
