+++ packages/server/src/tools/tools.test.ts
@@ -11,11 +11,6 @@ import { DEFAULT_CONFIG, type PolygConfig } from '@polyg-mcp/shared';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { PolygMCPServer } from '../server.js';
 
-// Helper to wait for FalkorDB eventual consistency
-// FalkorDB writes may not be immediately visible to subsequent queries
-const waitForConsistency = (ms = 500) =>
-  new Promise((resolve) => setTimeout(resolve, ms));
-
 // Test config with mock API key
 const TEST_CONFIG: PolygConfig = {
   ...DEFAULT_CONFIG,
@@ -108,24 +103,14 @@ describe('Tool Handlers', () => {
       expect(result.structuredContent).toBeDefined();
     });
 
-    // TODO: Flaky due to FalkorDB eventual consistency issues when running in parallel
-    it.skip('should include relationships when requested', async () => {
+    it('should include relationships when requested', async () => {
       const graphs = server.getOrchestrator().getGraphs();
-      const uniqueId = Date.now().toString();
-      const entity1 = await graphs.entity.addEntity(
-        `Entity1_rel_${uniqueId}`,
-        'Type',
-      );
-      const entity2 = await graphs.entity.addEntity(
-        `Entity2_rel_${uniqueId}`,
-        'Type',
-      );
-      await waitForConsistency();
+      const entity1 = await graphs.entity.addEntity('Entity1', 'Type');
+      const entity2 = await graphs.entity.addEntity('Entity2', 'Type');
       await graphs.entity.linkEntities(entity1.uuid, entity2.uuid, 'KNOWS');
-      await waitForConsistency();
 
       const result = await callTool(server, 'get_entity', {
-        name: `Entity1_rel_${uniqueId}`,
+        name: 'Entity1',
         include_relationships: true,
       });
